Handle mouse hover event in AbsListView
On Android 8.0 and 8.1 a change of ViewRootImpl exits touch mode on
mouse/stylus click. As a result, AbsListView will keep selection on
position 0, and will not action if click the list which can be seen
by scrolling to visible.

Add onHoverEvent to handle the mouse event and deal with selection.

Bug: https://issuetracker.google.com/issues/69774628
Test: Scroll a long listview to the end, check the last item is clickable.
Fixes: 7614cdf9 ("Exit touch mode on mouse/stylus click")
Tested-by: Chih-Wei Huang <cwhuang@linux.org.tw>
Change-Id: Ic6acc33779283342abc14d37a7d0514488a5cd1f