bluetooth: Fix for Can't enter "Connected devices" after pairing
Issue: Once the device is paired and connected the Connected devices
app is crashing.

Reason: When the Setting App is checking for the connected devices to
be updated on the UI, it is checking for the type of Audio profile i.e.
whether the profile is HEADSET or A2DP. Here we are getting the profile
as A2DP i.e. A2DP source profile And then checking for the A2DP
connection status. But in this case, BXT is acting as A2DP Sink device,
so its not able to find the reference for A2DP source profile, due to
which app is crashing.

Fix: Added a check for the null case, where it is checking if no
reference for A2DP source profile is found then don't check for its
connection status. Also, if no A2DP source profile is found, check if
A2DP sink profile is present. If present, then return the connection
status for A2DP sink profile otherwise return false.

Tracked-On: None

Signed-off-by: Gaganpreet kaur <gaganpreetx.kaur@intel.com>
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>