Support focused change betweeen launcher and freeform windows
If we start some freeform windows above the launcher, and click the empty
space of launcher, the launcher will be the focused task, and move to front
of the freeform windows. But we don't want to keep this behavior, and want
to launcher responses to input event, and keep the freeform windows above
the launcher. So we will do nothing if we found the new focused task is
the home type, we don't notify the AMS to change the focused task.
If we click another freeform window or normal fullscreen window, the logic
is the same as origin, the clicked freeform window or normal fullscreen window
will be the new focused task.

Note: the input response is processed by InputFlinger, see the InputDispatcher's
findTouchedWindowTargetsLocked to find the magic.

TODO: If we click the empty space of launcher, we should clear the focused
state of current freeform window.

Change-Id: Id2b6a8cd8b36ba23b3c3099c05a4a3d621207e43
Signed-off-by: utzcoz <utzcoz@outlook.com>