Use display bounds as freeform stack bounds
Many places use the stack bounds to clip the task or window bounds.
We use task bounds as stack bounds before, because every freeform window
uses a stack. But we don't consider shadow boundary size. To avoid
this problem, we use the display size as the stack size.

This commit will fix the shadow disappear after resizing and moving.

Change-Id: I6fe078ac8def31ed862b64dcf552e04ea4098751
Signed-off-by: utzcoz <utzcoz@outlook.com>