Set default refresh rate to 61Hz
The default refresh rate is used as an upper bound for the display mode
to select. However some monitors are using a frequency slightly larger
than 60Hz, while also supporting a mode with the same resolution, but
lower frequency (e.g. Dell Inspiron laptop). This makes Android to
change the display mode, which triggers what looks like a bug in the
kernel amdgpu driver, usually ending up in a kernel panic.

This change is a work around, to prevent Android from attempting to
change the refresh rate of the display (hence avoiding the kernel bug).

Change-Id: I9664422ba78949b383a34d661084c3aef9684f80